requirements.txt
asgiref==3.8.1
diskcache==5.6.3
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Jinja2==3.1.4
llama_cpp_python==0.3.5
MarkupSafe==3.0.2
numpy==2.2.0
pillow==11.0.0
PyJWT==2.10.1
rest-framework-simplejwt==0.0.2
shortuuid==1.0.13
sqlparse==0.5.3
typing_extensions==4.12.2
